move local modules to lib/
