+++ lib/normalize-api-path.js
